data_simulation_tool/requirements.txt
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
pandas==2.0.2
py2neo==2021.2.3
pyrootutils==1.0.4
PyYAML==6.0
retry==0.9.2
scipy==1.10.1
py7zr==0.20.6